requirements.txt
streamlit
groq
pandas
python-dotenv
mysql-connector-python